+++ modules/assays/backend/app/controllers/grit/assays/assay_data_sheet_definitions_controller.rb
@@ -20,6 +20,50 @@ module Grit::Assays
   class AssayDataSheetDefinitionsController < ApplicationController
     include Grit::Core::GritEntityController
 
+    def create_bulk
+      permitted = params.permit(sheets: [ *self.permitted_params, columns: [ :name, :safe_name, :description, :sort, :required, :data_type_id, :unit_id ] ])
+      errors = []
+      sheets = []
+      AssayDataSheetDefinition.transaction do
+        permitted[:sheets].each_with_index do |sheet, sheetIndex|
+          begin
+            columns = sheet[:columns]
+            sheet.delete(:columns)
+            assay_data_sheet_definition = AssayDataSheetDefinition.create(sheet)
+            sheets.push(assay_data_sheet_definition)
+            unless assay_data_sheet_definition.errors.blank?
+              assay_data_sheet_definition.errors.each do |e|
+                errors.push({message: e.message, path: ["sheets", sheetIndex, e.attribute]})
+              end
+            else
+              columns.each_with_index do |column, columnIndex|
+                assay_data_sheet_column = assay_data_sheet_definition.assay_data_sheet_columns.create(column)
+                if assay_data_sheet_column.errors
+                  assay_data_sheet_column.errors.each do |e|
+                    errors.push({message: e.message, path: ["sheets", sheetIndex, "columns", columnIndex, e.attribute]})
+                  end
+                end
+              end
+            end
+          rescue StandardError => e
+            logger.warn e.to_s;
+            logger.warn e.backtrace.join("\n");
+            errors["form"] ||= []
+            errors["form"].push e.to_s
+          end
+        end
+        unless errors.blank?
+          render json: { success: false, errors: errors }, status: :unprocessable_entity
+          raise ActiveRecord::Rollback
+        end
+        render json: { success: true, data: sheets }
+      end
+    rescue StandardError => e
+      logger.warn e.to_s;
+      logger.warn e.backtrace.join("\n");
+      render json: { success: false, errors: e.to_s }, status: :internal_server_error
+    end
+
     private
 
     def permitted_params