+++ modules/compounds/backend/app/models/grit/compounds/batch_property.rb
@@ -33,23 +33,15 @@ module Grit::Compounds
       update: [ "Administrator", "CompoundAdministrator" ],
       destroy: [ "Administrator", "CompoundAdministrator" ]
 
-    validates :safe_name, uniqueness: true, length: { minimum: 3 }
+    validates :safe_name, uniqueness: true, length: { minimum: 3, maximum: 30 }
     validates :safe_name, format: { with: /\A[a-z_]{2}/, message: "should start with two lowercase letters or underscores" }
     validates :safe_name, format: { with: /\A[a-z0-9_]*\z/, message: "should contain only lowercase letters, numbers and underscores" }
     validate :safe_name_not_conflict
 
     def safe_name_not_conflict
       return unless self.safe_name_changed?
-      if Grit::Compounds::Batch.entity_properties.find { |p| p[:name] == self.safe_name }
+      if Grit::Compounds::Batch.respond_to?(self.safe_name)
         errors.add("safe_name", "cannot be used as a safe name")
-      else
-        begin
-          # This is needed because of how active_support handles serialization as_json
-          Grit::Compounds::Batch.send(self.safe_name)
-        rescue NoMethodError => e
-        rescue StandardError => e
-          errors.add("safe_name", "cannot be used as a safe name")
-        end
       end
     end