+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.0] - 2025-08-01
+
+Assay Model Data Sheet loader
+
+### Added
+- (assays) assay model data sheet loader
+- (assays) clone assay models, data sheets, and columns
+
+### Fixed
+- (assays, compounds) crashes caused by too long safe names
+
+
 ## [0.7.0] - 2025-10-01
 
 Data tables